:alien: changing rating endpoint
