+++ lib/rest.js
@@ -30,12 +30,14 @@ const getData = (req, res, type) => {
  */
 // API
 const getGamesData = (req, res) => getData(req, res, 'games')
-const getMembersData = (req, res) => getData(req, res, 'members')
+// const getMembersData = (req, res) => getData(req, res, 'members') // old endpoint
+const getMembersData = (req, res) => axios.get('http://localhost:3002/rest/users').then(data => res.status(200).send(data.data)).catch(err => res.status(err.code).send(err))
 const getBlogData = (req, res) => getData(req, res, 'blog')
 const getEventsData = (req, res) => getData(req, res, 'events')
 const getPatronsData = (req, res) => axios.get('http://localhost:3002/rest/patrons').then(data => res.status(200).send(data.data)).catch(err => res.status(err.code).send(err))
 // DATA
-const getRatingData = (req, res) => getData(req, res, 'rating')
+// const getRatingData = (req, res) => getData(req, res, 'rating')
+const getRatingData = (req, res) => axios.get('http://localhost:3002/rest/special/rating').then(data => res.status(200).send(data.data)).catch(err => res.status(err.code).send(err))
 const getUpdateData = (req, res) => getData(req, res, 'update')
 
 // API
