+++ src/components/pages/Ranking/Ranking.js
@@ -32,7 +32,7 @@ class PageRanking extends React.Component {
                     <div className='page-description'>
                     <p>Ranking system utilizes the games' score system. Depending on the game's individual difficulty level, it is given one of { rating.length } possible marks:</p>
 				        <ul>
-                            { rating.map((r, rIndex) => <li key={ `r-${rIndex}` }><i className={ r.link } /> - worth { r.score } pts - { r.desc } </li>) }
+                            { rating.map((r, rIndex) => <li key={ `r-${rIndex}` }><i className={ r.icon } /> - worth { r.score } pts - { r.description } </li>) }
 				        </ul>
                     <p>Completing a game might mean earning its most demanding achievement, or getting the in-game 100%; but for the sake of simplicity the ranking system present here assumes that completing a game means earning 100% of its Steam achievements. You are awarded points depending on the completed game's difficulty level, which are later summarized and used to determine your placement on the ranking ladder.</p>
                     </div>
