+++ src/shared/helpers/helper.js
@@ -1,4 +1,4 @@
 export const swapRatingToIcon = (score, rating) =>
     rating
-        ? rating.find(r => r.score === score).link
+        ? rating.find(r => r.score === score).icon
         : 'fas fa-spinner'
