+++ src/shared/store/modules/Profiles.js
@@ -1,0 +1,10 @@
+// CONSTANTS
+export const SHOW_PROFILE = 'SHOW_PROFILE'
+
+// ACTION CREATORS
+export function showProfile(id) {
+    return {
+        type: SHOW_PROFILE,
+        id
+    }
+}
