+++ src/shared/store/store.js
@@ -5,6 +5,7 @@ import rating from '../config/rating.json'
 import { CHANGE_TAB } from './modules/Tabs'
 import { SEARCH_GAMES_VALUE, SEARCH_MEMBERS_VALUE } from './modules/Search'
 import { SHOW_GAMES_RATED } from './modules/CheckBoxes'
+import { SHOW_PROFILE } from './modules/Profiles'
 import { SHOW_LOGIN_MODAL, LOG_IN_USER, LOG_OUT_USER } from './modules/Login'
 import { CACHE_GAMES, CACHE_MEMBERS, CACHE_RATING, CACHE_EVENTS, CACHE_BLOG, CACHE_PATRONS } from './modules/Cache'
 
@@ -27,7 +28,8 @@ const defaultState = {
     events: [],
     blog: [],
     patrons: [],
-    rating: null
+    rating: null,
+    profileID: null
 }
 
 const enhancers = [ ]
@@ -60,6 +62,10 @@ const reducer = (state = defaultState, action) => {
         case SHOW_LOGIN_MODAL: return {
             ...state,
             showLoginModal: !state.showLoginModal
+        }
+        case SHOW_PROFILE: return {
+            ...state,
+            profileID: action.id
         }
         case LOG_IN_USER: return {
             ...state,
