+++ src/styles/sass/profile_old.sass
@@ -13,32 +13,6 @@
   width: 100%
   box-sizing: border-box
   background-color: $superlightgrey
-
-
-.profile-avatar 
-  width: 128px
-  min-width: 128px
-  height: 128px
-  min-height: 128px
-  margin: 15px
-  border: 5px solid #fff
-  background-color: #000
-  border-radius: 10px
-  box-sizing: border-box
-  box-shadow: 0 0 10px #000
-  padding: 2px
-  &.tier1
-    border-color: $tier1
-  &.tier2
-    border-color: $tier2
-  &.tier3
-    border-color: $tier3
-  &.tier4
-    border-color: $tier4
-.profile-patron
-  cursor: help
-
-
 .wrapper-profile > *
   width: 100%
 .profile-section
